refactor rules.go into separate cname and ns detection logic files
